+++ inst/app/ui.R
@@ -119,6 +119,7 @@ page_navbar(
     uiOutput("select_workpackages"),
     uiOutput("select_tasks"),
     uiOutput("select_expenses"),
+    checkboxInput("include_fte", "Include FTE positions?", value = TRUE),
     hr(),
     # selectInput("costing_type", "Costing type", c("CTU standard", "SNF")),
     HTML(